LayoutTests/inspector/profiler/memory-instrumentation-external-array.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="memory-instrumentation-test.js"></script>
<script>

var bigExternalArray = new Int32Array(1024 * 1024);

function test()
{
    var externalArraysExpectedMinimum = 4 * 1024 * 1024;
    InspectorTest.validateMemoryBlockSize(["ProcessPrivateMemory", "JSExternalResources"], externalArraysExpectedMinimum);
}
</script>
</head>

<body onload="runTest()">
<p>
This test checks that page's external arrays size reported by the memory agent
includes size of all external arrays created by the page.
<a href="https://bugs.webkit.org/show_bug.cgi?id=113790">Bug 113790.</a>
</p>
</body>
</html>
